Rebased with master and also fixed log4j issues
Rebased with master so we building on 0.5.0 release.
Also, fixed the issue with log4j issues with war file created. Currently
all the plugins are briging in their own log4j implementation when they
are imported to web project. I have excluded the providers from all of
them. This ensures that web is replying the plugins to provide its own
log4j implementation and not expect the web module to carry any.